# docs/swagger.yaml
definitions:
...
  dto.PeerListItemOutput:
    properties:
      create_at:
        type: string
      id:
        type: integer
      ip:
        type: string
      is_delete:
        type: integer
      name:
        type: string
      org:
        type: string
      port:
        type: integer
      state:
        type: string
      update_at:
        type: string
    type: object
  dto.PeerListOutput:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.PeerListItemOutput'
        type: array
      total:
        type: integer
    type: object
...
  middleware.Response:
    properties:
      data: {}
      errmsg:
        type: string
      errno:
        type: integer
      stack: {}
      trace_id: {}
    type: object
info:
  contact: {}
paths:
...
  /app/app_list:
    get:
      consumes:
      - application/json
      description: 节点列表
      operationId: /app/app_list
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页多少条
        in: query
        name: page_size
        required: true
        type: string
      - description: 页码
        in: query
        name: page_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PeerListOutput'
              type: object
      summary: 节点列表
      tags:
      - 节点管理
...
swagger: "2.0"